cleanup, removal of dead code, new file for thunks
